move detach to child class
