+++ matchzoo/preprocessor/dssm_preprocessor.py
@@ -3,6 +3,7 @@
 from matchzoo import engine
 from matchzoo import preprocessor
 from matchzoo import datapack
+from typing import Union, List, Tuple
 
 
 class DSSMPreprocessor(engine.BasePreprocessor):
@@ -89,3 +90,32 @@ class DSSMPreprocessor(engine.BasePreprocessor):
         if labels:
             data['labels'] = labels
         return datapack.DataPack(data=data, context=self._context)
+
+    def _detach_labels(
+        self,
+        inputs: list
+    ) -> Union[List[Tuple[str, str]], Union[list, None]]:
+        """
+        Detach labels from inputs.
+
+        During the training and testing phrase, :attr:`inputs`
+        usually have different formats. The training phrase
+        inputs should be list of triples like [(query, document, label)].
+        While during the testing phrase (predict unknown examples),
+        the inputs should be list of tuples like [(query, document)].
+        This method is used to infer the stage, if `label` exist, detach
+        label to a list.
+
+        :param inputs: List of text-left, text-right, label triples.
+        :return: Zipped list of text-left and text-right.
+        :return: Detached labels, list or None.
+        """
+        unzipped_inputs = list(zip(*inputs))
+        if len(unzipped_inputs) == 3:
+            # training stage.
+            return zip(unzipped_inputs[0],
+                       unzipped_inputs[1]), unzipped_inputs[2]
+        else:
+            # testing stage.
+            return zip(unzipped_inputs[0],
+                       unzipped_inputs[1]), None
